# Allow regex to fetch files from S3