--- a/cli/download_source.go
+++ b/cli/download_source.go
@@ -89,7 +89,7 @@ func downloadTerraformSourceIfNecessary(terraformSource *TerraformSource, terrag
 	}
 
 	terragruntOptions.Logger.Debugf("Downloading Terraform configurations from %s into %s", terraformSource.CanonicalSourceURL, terraformSource.DownloadDir)
-	if err := util.GetCopy(terraformSource.DownloadDir, terraformSource.CanonicalSourceURL.String()); err != nil {
+	if err := util.GetCopy(terraformSource.DownloadDir, terraformSource.CanonicalSourceURL.String(), ""); err != nil {
 		return err
 	}