--- a/util/get.go
+++ b/util/get.go
@@ -35,7 +35,7 @@ const (
 //
 // This copy will omit and dot-prefixed files (such as .git/, .hg/) and
 // can't be updated on its own.
-func GetCopy(dst, src string) error {
+func GetCopy(dst, src, fileRegex string) error {
 	// Create the temporary directory to do the real Get to
 	tmpDir, err := ioutil.TempDir("", "tf")
 	if err != nil {
@@ -45,8 +45,12 @@ func GetCopy(dst, src string) error {
 
 	tmpDir = filepath.Join(tmpDir, "module")
 
+	options := []getter.ClientOption{}
+	if fileRegex != "" {
+		options = append(options, getter.WithRegexFileMatcher(fileRegex))
+	}
 	// Get to that temporary dir
-	if err := getter.Get(tmpDir, src); err != nil {
+	if err := getter.Get(tmpDir, src, options...); err != nil {
 		return err
 	}