This tells go-getter to not fetch an entire directory (only files that match the regex)

Import benchmarks with three projects:
Small: 0.917s -> 0.357s
Medium: 2.82s -> 1.13s
Large: 22.1s -> 4.12s

The change to go-getter is this one: https://github.com/hashicorp/go-getter/pull/289
I'll wait for feedback but for the meantime, we can use the fork